My website's pics
You can see my website through the pictures. I am backend developer.  I have added login and register features to my website. My example website have blog on the navbar and it belongs to movie for the present and the blog have a category, author of movies. In this posses to search system for find to title and excerpt of movie. For the being present I can say that for frontend of my website works in progress. I hope you would like my works.